+++ CHANGELOG.md
@@ -7,5 +7,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## Changed
+#### Changed
 - Transform `npm-storecomponents` in a app that can be used in a VTEX store.
