+++ react/CollectionBadges.js
@@ -1,0 +1,3 @@
+import CollectionBadges from './components/collection-badges/src/index'
+
+export default CollectionBadges
