+++ react/ProductDescription.js
@@ -1,0 +1,3 @@
+import ProductDescription from './components/product-description/src/index'
+
+export default ProductDescription
